requirements.txt
Adafruit-Blinka==8.46.1
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.1
adafruit-circuitpython-framebuf==1.6.5
adafruit-circuitpython-requests==4.1.6
adafruit-circuitpython-ssd1306==2.12.17
adafruit-circuitpython-typing==1.10.3
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.72.1
Adafruit-PureIO==1.1.11
Adafruit-SSD1306==1.6.2
binho-host-adapter==0.1.6
colorzero==2.0
dukpy==0.4.0
gpiozero==2.0.1
mutf8==1.0.6
pillow==10.4.0
pyftdi==0.55.4
pyserial==3.5
pyusb==1.2.1
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
spidev==3.6
sysv_ipc==1.1.0
typing_extensions==4.12.2
you-get==0.4.1730
